Pulled thumbnail selection in separate component
